added begunstigde translation to lang files law.begunstige

+++ database/migrations/2023_06_29_155225_alter_table_invoice_requests_add_user_id_field.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::table('invoice_requests', function (Blueprint $table) {
 
-            $table->unsignedBigInteger('user_id')->unsigned();
+            $table->BigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
